Refactor ConsoleCommand utility class
